moving non-views out of views into utils
